[CC-9] Share logic using KMM
- Move `Game` to shared module.
- Adapted for android and iOS.

+++ iosApp/iosApp/IncrementalText.swift
@@ -1,7 +1,7 @@
 import SwiftUI
 
 struct IncrementalText: View {
-    var value: Int
+    var value: Int64
     
     var body: some View {
         VStack {
